tracking/requirements.txt
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.4
psycopg2==2.8.6
requests~=2.22.0
pyyaml~=5.3.1
coreapi==2.3.3